testing stop script

+++ codedeploy/ApplicationStop.sh
@@ -1,2 +1,9 @@
 #!/bin/bash
-sudo killall --signal SIGINT griddy
+
+if pgrep -x "griddy" > /dev/null
+then 
+  echo "Running"
+  sudo killall --signal SIGINT griddy
+else 
+  echo "Sleeping"
+fi
